style: moved specs to separate folder near tested files
