+++ packages/core/src/error/specs/error.effect.spec.ts
@@ -1,8 +1,8 @@
 import { mapTo } from 'rxjs/operators';
-import { Marbles } from '../../../util/marbles.spec-util';
-import { HttpRequest, HttpResponse } from '../http.interface';
-import { error$, errorEffectProvider } from './error.effect';
-import { HttpError } from './error.model';
+import { Marbles } from '../../../../util/marbles.spec-util';
+import { HttpRequest, HttpResponse } from '../../http.interface';
+import { error$, errorEffectProvider } from '../error.effect';
+import { HttpError } from '../error.model';
 
 const createMockRes = () => ({} as HttpResponse);
 const createMockReq = (url = '/') => ({ url } as HttpRequest);
