+++ packages/core/src/router/specs/urlParams.factory.spec.ts
@@ -1,4 +1,4 @@
-import { factorizeRegExpWithParams } from './urlParams.factory';
+import { factorizeRegExpWithParams } from '../urlParams.factory';
 
 describe('Url parameters factory', () => {
